+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## development
+- Add compatibility with Grafana 8
 
 - Add Makefile to assist in packaging and publishing
 - Overlay image can now be refreshed with the dashboard. Thanks, @jreyesr!